+++ src/hooks/useConfirm.ts
@@ -1,15 +1,8 @@
 import { useState } from 'react';
 
-interface useConfirmProps {
-  openConfirm: () => void;
-  closeConfirm: () => void;
-  handleConfirm: () => void;
-  isOpen: boolean;
-}
-
 // custom window.confirm modal 사용하는 페이지에 사용할 훅
 
-const useConfirm = (onConfirm: () => void): useConfirmProps => {
+const useConfirm = (onConfirm: () => void) => {
   const [isOpen, setIsOpen] = useState(false);
 
   const openConfirm = () => {
