+++ src/hooks/useScrollMemory.ts
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 // 뒤로 가기 등 타 페이지 방문 후 다시 돌아왔을 때 마지막 스크롤 위치 기억해야하는 페이지에서 사용할 훅
 
-const useScrollMemory = () => {
+const useMemoScrollPosition = () => {
   // 컴포넌트 마운트 시 이전 스크롤 위치를 기억해 이동하는 useEffect
   useEffect(() => {
     const lastScrollTop = Number(sessionStorage.getItem('scroll_top'));
@@ -19,4 +19,4 @@ const useScrollMemory = () => {
   }, []);
 };
 
-export default useScrollMemory;
+export default useMemoScrollPosition;
